+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Extended support for description on frontpage, with corpus updates [#341](https://github.com/spraakbanken/korp-frontend/issues/341)
 - The "medial part" option now includes first/last parts also for lemgram search [#347](https://github.com/spraakbanken/korp-frontend/issues/347)
+- Improved UI reactivity for Simple search
 - Removed globals `settings`, `currentMode`, `currentModeParallel`, `model`, `CorpusListing`, `ParallelCorpusListing` – import them instead
 
 ### Fixed